requirements.txt
pytest==7.4.0
requests==2.31.0
numpy>=1.26,<2
openai==1.79.0
anthropic==0.25.0
autoflake==2.1.1
flake8==7.3.0
flake8-import-order==0.19.2
black==23.9.1
isort==5.12.0
pylint==3.2.6
mypy==1.11.2
pyflakes==3.4.0
pycodestyle==2.14.0
typing_extensions==4.13.1
types-requests==2.28.11.17
pyright==1.1.402
pydantic==2.7.1
google-genai==1.24.0
